Move db row representations to iml-wire-types.
Db row structs are used in multiple crates. As such, they should live in
iml-wire-types. They also sometimes need a conversion from a postgres
`Row`, so we add a new feature `postgres-interop`, which when activated
provides the conversion.
