+++ iml-wasm-components/iml-fs/Cargo.toml
@@ -17,7 +17,7 @@ iml-paging = { path = "../iml-paging", version = "0.1" }
 iml-pie-chart = { path = "../iml-pie-chart", version = "0.1" }
 iml-tooltip = { path = "../iml-tooltip", version = "0.1" }
 iml-utils = { path = "../iml-utils", version = "0.1" }
-iml-wire-types = "0.1.10"
+iml-wire-types = "0.1"
 iml-stratagem = { path = "../iml-stratagem", version = "0.1" }
 log = "0.4"
 seed = "=0.4.0"
