+++ iml-wasm-components/iml-stratagem/Cargo.toml
@@ -17,19 +17,19 @@ iml-sleep = { path = "../iml-sleep", version = "0.1" }
 iml-toggle = { path = "../iml-toggle", version = "0.1.0" }
 iml-tooltip = { path = "../iml-tooltip", version = "0.1.0" }
 iml-utils = { path = "../iml-utils", version = "0.1" }
-iml-wire-types = "0.1.10"
+iml-wire-types = "0.1"
 iml-grafana-chart = { path = "../iml-grafana-chart", version = "0.1" }
 log = "0.4"
 seed = "=0.4.0"
 wasm-bindgen = { version = "0.2", features = ["serde-serialize"]}
 web-sys = "0.3"
-chrono = "0.4.7"
+chrono = "0.4.9"
 
 [dev-dependencies]
 wasm-bindgen-test = "0.2"
 
 [dev-dependencies.web-sys]
-version = "0.3.25"
+version = "0.3.27"
 features = [
   "Response"
 ]
